Refine the ManifestBitmap API for manifest deletion vectors (#7591)
#### Which Delta project/connector is this regarding?

- [x] Spark
- [ ] Standalone
- [ ] Flink
- [ ] Kernel
- [ ] Other (fill in here)

## Description

Refines the `ManifestBitmap` abstraction used for manifest
deletion-vector bitmaps.

`RoaringManifestBitmap` is now backed by an `Either[Array[Byte],
Seq[Int]]`, so a bitmap can be
constructed either from raw positions (`fromPositions`) or from
already-serialized bytes
(`fromSerializedByteArray`). Decoding into a `RoaringBitmapArray` is
deferred until the first
bit-level access, so a bitmap that is only carried forward and
re-serialized never pays the
decode/encode cost.

`serializeAsByteArray` returns the original bytes when the bitmap was
never mutated -- with a
defensive copy so a caller cannot mutate the internal buffer -- and
re-encodes from the decoded
bitmap once a position has been added. `BackReference.pos` is narrowed
from `Long` to `Int`, and
the manifest write path (`AMTWriteHelper`, `IncrementalAMTWriter`) and
the associated tests are
updated to the new API.

## How was this patch tested?

Updated the existing manifest, back-reference, and action-serializer
unit tests to exercise the
new API: `ActionSerializerSuite`, `AMTBackReferenceSuite`,
`AMTIncrementalWriteInvariantSuite`,
`AMTIncrementalWriteTestBase`, and `AMTSnapshotSuite`.

## Does this PR introduce _any_ user-facing changes?

No.

+++ spark/src/main/scala/org/apache/spark/sql/delta/amt/AMTCheckpointProvider.scala
@@ -169,7 +169,7 @@ final class AMTCheckpointProvider(
       .where(col("entry.tracking.status").isin(Tracking.Status.liveEntryStatuses.toSeq: _*))
       .filter { entryWithLoc =>
         mdvBroadcast.value.get(entryWithLoc.leafPath)
-          .forall(bytes => !AMTUtils.deserializeMdv(bytes).contains(entryWithLoc.pos))
+          .forall(bytes => !AMTUtils.deserializeMdv(bytes).contains(entryWithLoc.pos.toInt))
       }
 
     dataEntries
@@ -184,7 +184,7 @@ final class AMTCheckpointProvider(
                 val absLeaf = SparkPath.fromUrlString(entryWithLoc.leafPath).toPath
                 val relManifest =
                   AMTUtils.relativizeManifestPathToTableRoot(fs, localTableRoot, absLeaf)
-                Some(BackReference(relManifest, entryWithLoc.pos))
+                Some(BackReference(relManifest, entryWithLoc.pos.toInt))
               }
               // Root entries have nothing above them to inherit from, so they resolve against an
               // empty parent.
